Enabled gcc's -Wall option. Corrected all but one of the resulting warnings (see next commit).
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1219 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/WithinHost/Infection/MolineauxInfection.h
@@ -56,9 +56,9 @@ private:
     void updateGrowthRateMultiplier();
 
 	// v: number of variants per clone (one infection = one new clone)
-	static const int v = 50;
+	static const size_t v = 50;
 	// taus: used for the variantTranscending and variantSpecific array, 4 Molineaux timesteps = 8 days
-	static const int taus = 4;
+	static const size_t taus = 4;
 
 	///@brief static variables red from parameters
 	//@{
